situacion previa refactor

+++ suma_dos_numeros/suma_dos_numero.py
@@ -3,8 +3,8 @@ class SumaDosNumeros:
 
     def suma(a, b):
         try:
-          a = float(a)
-          b = float(b)
+          a = int(a)
+          b = int(b)
           return a + b 
         except:
            raise ValueError("Valor no valido")
